requirements.txt
django==2.2.13
djangorestframework==3.11.0
django-storages[google]
gunicorn==20.0.4

numpy
matplotlib
pandas
firebase-admin
python-dotenv

grpcio
google-api-core
google-cloud-core
google-cloud-bigquery==1.23.0
google-cloud-bigquery-storage==0.7.0
google-cloud-storage==1.28.0
google-cloud-datastore==1.12.0
google-cloud-pubsub